Moved accesspoint and ssid classes to its own file.
